+++ .gitignore
@@ -1,3 +1,3 @@
 .idea/*
-__pycache__/*
+**/__pycache__/*
 travel_agent_env/*
