--- Aufgaben/d_Agenten/b. Agent Loop/tool_determination.py
@@ -1,5 +1,0 @@
-def determine_tool(next_step, tools):
-    # use the given string to let an llm decide which tools to use
-    # return the choices[0].message.tool_calls[0] object
-
-# TODO write a dummy method that uses tools and the next_step "calculate the costs for a hotel with checkin date 27.03.2024, checkout date 10.04.2024 and a nightly cost of 235"
