+++ Aufgaben/d_Agenten/b. Agent/agent.py
