+++ Aufgaben/d_Agenten/b. Agent/exampleNextStep.py
@@ -1,0 +1,12 @@
+next_step = """
+    {
+        "chainOfThoughts": "The plan outlines four steps, and none have been executed yet. We should start by discovering the available hotels on Mars.",
+        "nextStep": {
+            "recommendedTool": "functions.list_hotels",
+            "Input": {
+                "location": "Mars"
+            }
+        },
+        "finish": false
+    }
+    """
