+++ Solutions/d_Agenten/b_Agent/tool_determination.py
@@ -1,0 +1,24 @@
+from openai import OpenAI, OpenAI
+from Solutions.d_Agenten.tools_description import tool_belt
+
+
+client = OpenAI()
+
+def determine_tool(next_step, tools):
+    completion_tool = client.chat.completions.create(
+        model="gpt-4o-mini",
+        tools=tools,
+        tool_choice='required',
+        messages=[
+            {"role": "system",
+             "content": """Read the assistant message below and determine which tool to use to satisfy this task"""},
+            {
+                "role": "assistant",
+                "content": next_step
+                }
+            ]
+        )
+    return completion_tool.choices[0]
+
+if __name__ == "__main__":
+    print(determine_tool("calculate the costs for a hotel with checkin date 27.03.2024, checkout date 10.04.2024 and a nightly cost of 235", tool_belt))
